# frontend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Frontend",
        "version": "v0.0.0"
    },
    "paths": {
...
        "/api/frontend/v1/websites": {
            "get": {
                "tags": [
                    "Websites"
                ],
                "summary": "Retrieve all websites owned by this user",
                "description": "Retrieve all websites owned by this user",
                "operationId": "get_all_websites",
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListWebsites"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Websites"
                ],
                "summary": "Create a new website",
                "description": "Create a new website",
                "operationId": "create_website",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateWebsiteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UuidSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/websites/{uuid}": {
            "get": {
                "tags": [
                    "Websites"
                ],
                "summary": "Retrieve a single website",
                "description": "Retrieve a single website",
                "operationId": "get_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "description": "The uuid",
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FullWebsite"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Websites"
                ],
                "summary": "Update a website",
                "description": "Update a website",
                "operationId": "update_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "description": "The uuid",
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWebsiteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Empty body"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Websites"
                ],
                "summary": "Delete a website",
                "description": "Delete a website",
                "operationId": "delete_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "description": "The uuid",
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Empty body"
                    }
                }
            }
        },
        "/api/frontend/v1/websites/{uuid}/domains": {
            "post": {
                "tags": [
                    "Websites"
                ],
                "summary": "Add a domain to a website",
                "description": "Add a domain to a website",
                "operationId": "add_domain_to_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "description": "The uuid",
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddDomainToWebsiteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UuidSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/websites/{website_uuid}/domains/{domain_uuid}": {
            "delete": {
                "tags": [
                    "Websites"
                ],
                "summary": "Remove a domain from a website",
                "description": "Remove a domain from a website",
                "operationId": "remove_domain_from_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "domain_uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    },
                    {
                        "in": "path",
                        "name": "website_uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Empty body"
                    }
                }
            }
        },
        "/api/frontend/v1/websites/{uuid}/deploy": {
            "post": {
                "tags": [
                    "Websites"
                ],
                "summary": "Deploy the configuration to the webserver.",
                "description": "Deploy the configuration to the webserver.\n\nThis will configure the webspace and request certificates for all added domains\n\nReturns an uuid that will be used to send a notification via websocket when the deployment\nprocess has finished",
                "operationId": "deploy_website",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "description": "The uuid",
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UuidSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/websites/{uuid}/check-dns": {
            "post": {
                "tags": [
                    "Websites"
                ],
                "operationId": "check_dns",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UuidSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "description": "The response that is sent in a case of an error",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "A human-readable error message.\n\nMay be used for displaying purposes",
                        "type": "string"
                    },
                    "status_code": {
                        "$ref": "#/components/schemas/ApiStatusCode"
                    }
                },
                "required": [
                    "message",
                    "status_code"
                ]
            },
...
            "CreateWebsiteRequest": {
                "description": "The request to create a website",
                "type": "object",
                "properties": {
                    "name": {
                        "description": "The name of the website",
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ]
            },
...
            "FullWebsite": {
                "description": "The full representation of a website",
                "type": "object",
                "properties": {
                    "created_at": {
                        "$ref": "#/components/schemas/SchemaDateTime"
                    },
                    "deploy_state": {
                        "$ref": "#/components/schemas/DeployState"
                    },
                    "domains": {
                        "description": "A list of domains for this website",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FullWebsiteDomain"
                        }
                    },
                    "last_deployment": {
                        "$ref": "#/components/schemas/SchemaDateTime"
                    },
                    "name": {
                        "description": "Descriptive name of the website",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "The unique key of a website",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "created_at",
                    "deploy_state",
                    "domains",
                    "name",
                    "uuid"
                ]
            },
...
            "ListWebsites": {
                "description": "A list of websites",
                "type": "object",
                "properties": {
                    "websites": {
                        "description": "The list of websites",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SimpleWebsite"
                        }
                    }
                },
                "required": [
                    "websites"
                ]
            },
...
            "UuidSchema": {
                "description": "a uuid",
                "type": "object",
                "properties": {
                    "uuid": {
                        "description": "The uuid",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "uuid"
                ]
            },
...
        }
    }
}